change type of wage parameter for setWage setter

+++ crud-java-oracle/src/br/com/fiap/mercado/view/Employee.java
@@ -41,7 +41,7 @@ public class Employee {
 	public void setEmail(String email) {
 		this.email = email;
 	}
-	public void setWage(float wage) {
+	public void setWage(double wage) {
 		this.wage = wage;
 	}
 }
